move Main.js to common template
